endpoint fixes
